requirements.txt
pyspark==3.3.0
kafka-python==2.0.2
pandas
boto3
apache-airflow